loadProcess uses a parameter now to connect to different Servers and from to can also be passed as an argument. Added Tests for MockMotor and MockSensor

+++ MES/plc/javafunctionalunit/src/main/java/hardware/actuators/Motor.java
@@ -30,5 +30,5 @@ public abstract class Motor {
      * Waits x milliseconds. If motor library has a delay method, use it instead of thread.sleep
      * @param msDelay delay in ms
      */
-    public abstract void waitMs(int msDelay);
+    public abstract void waitMs(long msDelay);
 }
